% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeAPIrequester.R
\name{initializeAPIrequester}
\alias{initializeAPIrequester}
\title{Initialize the API requester}
\usage{
initializeAPIrequester(appName, apiKeys)
}
\arguments{
\item{appName}{The name of your app}

\item{apiKeys}{The API keys to be used for the pooling of API keys in apiRequest}
}
\value{
TRUE, if initialization was successful.
}
\description{
Used to set the keys to be used for the API key pooling. This function also sets
the name of your app, which is passed into the API as a comment and visible to the
owners of the API keys used in their API settings page.
}
