% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpooledAPIrequest.R
\name{unpooledAPIRequest}
\alias{unpooledAPIRequest}
\title{unpooledAPIRequest}
\usage{
unpooledAPIRequest(
  keyToUse,
  section,
  selection,
  IDtoSend = "",
  comment = "",
  timestampToSend = "",
  personalStatsToRequest = c(),
  errorBehaviour = "silent"
)
}
\arguments{
\item{keyToUse}{The API key to use for the request.}

\item{section}{The section component of your request.}

\item{selection}{The selection component of your request}

\item{IDtoSend}{The ID to send with the request. Optional, defaults to no ID.}

\item{comment}{Any additional comment to send with your request. Will always be prefixed with your
app name, which you set with initializeAPIrequester. Optional, if omitted, comment will only be your
app name.}

\item{timestampToSend}{If you are requesting past personal stats data, this is where you set the time
from which you want the data.}

\item{personalStatsToRequest}{If requesting past personal stats data, this is where you specify the stats
you want. Expected as a vector of stats. Note that the API does not permit more than ten stats at a time.}

\item{errorBehaviour}{}
}
\value{
The response of the API.
}
\description{
Sends a single request to the Torn API using a specified key, instead of the pooled
behaviour of the apiRequest function. This is intended for when you are requesting a
specific user's data, and so really need to be using their specific key - for a request
that can be made with a public access API key, you are better off using the apiRequest
function.
}
\details{
This function implements all the safety checks the package provides - as far as possible
it checks that your input makes sense for the request you are making, and it also looks at
what the API returns and throws errors if you got an API error that needs handling, such
as your key being invalid or that you have been rate limited.
}
